+++ src/public/templates/alma-checkout-plans-classic.php
@@ -1,0 +1,14 @@
+<?php
+/**
+ * Template.
+ *
+ * @since 4.2.0
+ *
+ * @package Alma_Gateway_For_Woocommerce
+ * @subpackage Alma_Gateway_For_Woocommerce/public/templates
+ */
+
+?>
+<div>
+<?php echo $description; // phpcs:ignore WordPress.Security.EscapeOutput ?>
+<br>