+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+v5.0.0
+------
+* feat: In Page checkout
+* feat: Separate the gateways in the checkout page
+* feat: Compatibility Woocommerce 8.0.1
+* feat: Compatibility Woocommerce 8.0.2
+* feat: Compatibility WordPress 6.3
+
 v4.3.4
 ------
 * fix: wrong variable in mismatch